Remove React fixture tsconfig
