+++ packages/genui/ui-judge/tests/fixtures/react/lynx.config.js
@@ -5,6 +5,9 @@ import { pluginReactLynx } from '@lynx-js/react-rsbuild-plugin';
 import { defineConfig } from '@lynx-js/rspeedy';
 
 export default defineConfig({
+  source: {
+    entry: './src/index.jsx',
+  },
   plugins: [
     pluginReactLynx(),
   ],
