--- packages/genui/ui-judge/tests/fixtures/react/src/assets.d.ts
@@ -1,8 +1,0 @@
-// Copyright 2026 The Lynx Authors. All rights reserved.
-// Licensed under the Apache License Version 2.0 that can be found in the
-// LICENSE file in the root directory of this source tree.
-
-declare module '*.png' {
-  const src: string;
-  export default src;
-}
