--- packages/genui/ui-judge/tests/fixtures/react/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "../../../../../../tsconfig.json",
-  "compilerOptions": {
-    "jsx": "react-jsx",
-    "jsxImportSource": "@lynx-js/react",
-    "noEmit": true,
-
-    "allowJs": true,
-    "checkJs": true,
-    "isolatedDeclarations": false,
-  },
-  "include": ["src", "lynx.config.js"],
-  "references": [
-    { "path": "../../../../../react/tsconfig.json" },
-    { "path": "../../../../../rspeedy/core/tsconfig.build.json" },
-    { "path": "../../../../../rspeedy/plugin-react/tsconfig.build.json" },
-  ],
-}
